compile to .pyd
